fix: print_bytes() loop counter size_t to match pos
Copilot re-review on PR #386 (low-confidence note): now that `pos` is
`size_t`, the deprecated debug helper `print_bytes()` should use `size_t`
for its loop index too, otherwise a codestream > 4 GiB would overflow
the uint32_t counter and cause an infinite loop.  Dead code path
(MAYBE_UNUSED + deprecated) but worth keeping consistent.

+++ source/core/codestream/codestream_destination.cpp
@@ -91,7 +91,7 @@ int32_t j2c_dst_memory::flush(std::vector<uint8_t> *obuf) {
 size_t j2c_dst_memory::get_length() const { return pos; }
 
 OPENHTJ2K_MAYBE_UNUSED [[deprecated]] void j2c_dst_memory::print_bytes() {
-  for (uint32_t i = 0; i < pos; i++) {
+  for (size_t i = 0; i < pos; i++) {
     if (i % 32 == 0) {
       printf("\n");
     }
